src/requirements.txt
httpx==0.23.3
python-dotenv==0.21.0
swibots==1.0.2
websocket==0.2.1
websockets==10.4
